requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
cached-property==1.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
commonmark==0.9.1
dictdiffer==0.9.0
idna==3.6
notion==0.0.23
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.31.0
soupsieve==2.5
text-unidecode==1.3
tzlocal==5.2
urllib3==2.1.0
